[sil-combine] When simplifying convert functions, base whether or not an apply has nothrow on the actual function_ref.
Previously, we based it off of whether or not the original apply had a nothrow
bit. This is incorrect in the case where we added an error result to a function
without an error result. In such a case, we need to /not/ put on the nothrow
bit. This commit generalizes this idea slightly by assuming that if we are asked
to perform this transformation we should just match what the underlying
function_ref (i.e. setting nothrow if the underlying function type has an error
result and not setting nothrow if the underlying function type does not have an
error result).

rdar://47828439
(cherry picked from commit 4d90cddd2ebdcc283e81950f2fb2eb0d4c4ff7ab)