[update-checkout] Add new branch configs for Swift 5.1
Add a tentative branch config for "swift-5.1-branch". Not all of the repos
are branched yet. I went ahead and added a reference to swift-5.1-branch for
the Swift repo, so that I can use this to start pulling together something
with cherry-picks from master-next. (The current master-next branch has
moved past the point where we branched the LLVM repos for swift-5.1-branch,
so we can't just grab a copy of master-next.) I also added a separate config
for "swift-5.1-old-llvm-branch", which will be used temporarily to track the
same Swift content as swift-5.1-branch but still building with the version
of LLVM in swift-5.0-branch.