[AutoDiff] Release unneeded superset pullback results.
When differentiating a function that is known `@differentiable` with respect to a superset of parameters than needed for this differentiation, AD will use the superset version and use a subset of what its pullback returns. However, we forgot to release the pullback results that are unneeded, which caused memory leaks. This PR fixes that.

This PR also adds code that correctly handles indirect results.