MetadataReader: Fix caching of Objective C class metadata
If resolving the type of an instance produces a class metadata for
which we cannot build a type (for example, a special class like
__NSCFNumber, which the ClangImporter does not produce a ClassDecl
for), we try the superclass.

The caching logic was broken in this case, so subsequent calls
would return an empty type.

(cherry picked from commit 28f305bf75faed937fdca155b2f2e03d5febe1d9)