[sil] Tighten up SIL verification around when checked_cast_br should/shouldn't have failure arguments.
checked_cast_br in ownership SIL funnels through the original value to the
failure block as a SILArgument for cleanup purposes. In contrast, when SIL is
not in ossa, we do not want the failure bb case to have any arguments since that
is the pattern that SILOptimizer passes expect.

This commit just formalizes these constraints into the SILVerifier to catch
mistakes.