IRGen: followup for SVN r352827
This adjusts the previous change for the IRGen API changes in SVN r352827.  We
sometimes may get back a BitCastInst for the function type coercion.  Ensure
that we strip the pointer casts before performing the cast which may otherwise
fail with an assertions build.  We need to now perform a cast for the
existential boxed constructor and destructor calls to account for the missing
bitcast on a reused function.  This actually makes the IR a bit clearer as the
two parameters are being cast instead when necessary.