Assure DIFiles get the directory name during IRCodeGen
PDB generation assumes DIFiles have both a filename and a directory.
IRGenDebugInfoImpl currently checks if an absolute path is given for
the Filename and if not it just moves forward with only the filename
which causes PDB generation to be broken.

Clang handles a non absolute path for a file by by first querying the
DebugCompilationDir then the IGM.CWDName and lastly just getting
llvm::sys::fs::current_path. We should do the same here.