[AutoDiff] TF-235: Fix/simplify AdjointGen substitution map remapping.
The old substitution map remapping logic was incorrect and needlessly complex.
The new logic is correct and simpler.
There is certainly more room for simplification: to be explored.

Enables (partially correct) differentiation of `applied(to:)` for
generic layer combinators.

Partially resolves TF-235.
TF-242 will soon be resolved to fix `applied(to:)` adjoint segfault.