[AutoDiff] Work around SR-9595 and simplify `Differentiable` conformances.
The `Differentiable` protocol generic signature minimization bug (SR-9595)
has caused pain to all users of the `Differentiable` protocol conformances
(derived or not derived), and triggered crashers in nested differentiation.

This PR is a workaround that does the following:
- Moves all `Differentiable` requirements except recursive generic constraints
  to the `__Differentiable` and `_Differentiable` protocols.
- Makes `Differentiable` refine `_Differentiable` and specifies recursive
  constraints on `TangentVector` and `CotangentVector`. This enables the type
  checker and subsequent compilation phases to find `AdditiveArithmetic`
  conformances for `TangentVector` and `CotangentVector`.

`__Differentiable` and `_Differentiable` are internal protocols that are never
supposed to be used directly. This causes a number of minor issues, none of which
are worse than the original issue (requiring explicit constraints to
`AdditiveArithmetic` everywhere):
- User-facing issue: When a type does not fulfill protocol requirements to conform,
  the error shows "does not conform to `__Differentiable`" and "does not conform to
  `_Differentiable`", in addition to "does not conform to `Differentiable`". This
  is acceptable at the moment.
- Compiler issue: In principle, implied conformance is disallowed in conditional
  conformances. However, we do not want the user to conform their types to
  `_Differentiable` explicitly in addition to `Differentiable`, therefore special
  cases are added in the conformance checker to allow these implied conformances.

Partially resolves TF-213 by casting a user-facing issue into a compiler special
case. We expect to help fix SR-9595 ASAP, so these hacks will go away soon.

Test notes:
- All test changes were expected.
- `swift/test/AutoDiff/bad_reabstraction.swift` is no longer reproducible and
  has been deleted.