[AutoDiff] Fix memory leak on adjoint accumulation. (#22687)
Direct adjoint accumulation produces a new value, but we didn't create a cleanup for this new value, which caused a memory leak whenever there's a fanout in the original program. This PR fixes that.

We should add reference counting codegen tests systematically. This is fixing a P0, so I'm deferring that.