[AutoDiff] TF-246: Fix adjoint buffer calculation for projections.
Previously, adjoint buffers for projection instructions were newly
allocated buffers in the adjoint - this led to some derivative
miscalculations.

Now, adjoint buffers for address-to-address projections are projections
in the underlying buffer's adjoint buffer. This is semantically correct
and fixes previous miscalculations.

With this change, `AdjointEmitter` visitors for projections are no
longer necessary.

Change usefulness propagation through buffers to be recursive, instead
of stopping after one level.

Resolves TF-246.