[AST] NFC: Rename get*OrBoundGeneric* to be clearer and shorter
The get*OrBoundGeneric* methods are, in practice, simple getters that
extract a Decl node. Let's rename the APIs accordingly and make them
shorter at the same time. This is similar to the DeclContext changes
made by Jordan Rose in: 537954fb9338951ce94dbe8549bbc7f0e2b1395f

- getStructOrBoundGenericStruct -> getStructDecl
- getClassOrBoundGenericClass -> getClassDecl
- getEnumOrBoundGenericEnum -> getEnumDecl
- getNominalOrBoundGenericNominal -> getNominalTypeDecl
- getAnyNominal -> getNominalTypeDecl
- getAnyGeneric -> getGenericTypeDecl

This is mostly a mechanical change. That being said, getAnyNominal and
getNominalOrBoundGenericNominal don't behave exactly the same. The
former includes unbound generics whereas the latter does not.
Nevertheless, only one line of code needed to change as a consequence in
`OverrideFilteringConsumer::foundDecl()` (in LookupVisibleDecls.cpp).