[AutoDiff] TF-277: Fix type-remapping crash. (#22851)
* [AutoDiff] TF-277: Fix type-remapping crash.

Previously, `getOpASTType` was used in `visitLoadInst` and `visitStoreInst` to
remap SIL types that were converted to canonical types. However, this triggered
an assertion failure regarding remapping lowered types.

Now, `PrimalGenCloner::visitLoadInst` and `PrimalGenCloner::visitStoreInst` are removed, because they weren't supposed to be here in the first place if there were no bugs with the cloner.